Fix all compiler warnings; complete the WM IPC command set
Warning fixes (all real code changes, no suppressions):
- vantage-renderer: printf had 8 conversions but 7 args — vendor/driver/
  path columns were shifted and the last %s read a missing argument (UB);
  supply the vendor string for the vendor= field
- vt-theme: GTK3 CSS template had 25 conversions vs 26 args (one extra
  t->radius); drop it. Explicit float casts in hex->rgba (unsigned->float)
- vt-wallpaper: explicit (float) casts in gradient interpolation
- vt-wm-x11: _layer_of takes const _client_t*; remove dead _find_model
- vt-gpu: wire _device_name_from_uevent as device_str fallback when
  subsystem_device is absent
- vt-ipc: log handler found/none instead of casting a function pointer
  to void* (ISO C violation)
- vt-integ-audio: split double-if on one line (misleading indentation)
- vt-image, vt-desktop: reword comments containing '/*' sequences
- vt-wayland-testclient: guard _GNU_SOURCE against the global -D
- vt-backend-wayland: replace 7 incompatible function-pointer casts in
  interface structs with exact-signature handlers; drop unused keymap var

Functional completion found while validating (unused handlers were real
features, not dead code):
- New IPC message types WM_RESTORE/WM_UNMAXIMIZE/WM_UNFULLSCR/WM_WS_MOVE
  and SESSION_STATUS; register the previously-unregistered handlers
- vantage-remote gains restore/unmaximize/unfullscreen/resize/ws-move/
  status commands; maximize/fullscreen now truly toggle (as documented)

Bug fixes uncovered by the new commands:
- UnmapNotify double delivery: the WM selects StructureNotifyMask on the
  window AND SubstructureNotifyMask on the root, so each unmap arrives
  twice; the expect_unmap guard was consumed by the first copy and the
  second UNMANAGED the window — minimize and workspace moves made
  windows vanish from the model. Only the root variant is acted on now.
- Handler errors now return an IPC ERROR frame (VT_IPC_E_HANDLER) so
  clients fail fast with a clear message instead of a 2s timeout

+++ src/wallpaper/vt-wallpaper.c
@@ -271,7 +271,7 @@ void vt_wallpaper_render(vt_wallpaper_t *w, vt_renderer_t *r, vt_rect_t area) {
         /* simple horizontal/vertical gradient via N solid stripes */
         int stripes = 32;
         for (int i = 0; i < stripes; i++) {
-            float t = (float)i / stripes;
+            float t = (float)i / (float)stripes;
             vt_color_t c = {
                 w->color_a.r + (w->color_b.r - w->color_a.r) * t,
                 w->color_a.g + (w->color_b.g - w->color_a.g) * t,
